# crates/api/assets/openapi.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "SoraMetrics API",
  "version": "33.0.0",
  "summary": "Read-only analytics of the SORA v2 blockchain.",
  "description": "Public, read-only, no authentication. Data is indexed from the SORA v2 chain (Substrate) and, for balances, holders, pools, staking and governance, read from a node at request time.\n\nPrices are marginal on-chain quotes (`liquidityProxy.quote` of 0.000001 token to DAI, DAI = 1 USD). `illiquid: true` means selling 1 USD worth of the token returns under 0.50 USD: never derive a market cap or a holding value from that price.\n\nA swap is worth its cheaper priced leg (a marginal quote can overvalue the illiquid side by orders of magnitude). Swap rows carry that value in both `in.usd` and `out.usd`; swap volume, trends, top pair and accumulation are sums of it.\n\nToken supply shown by the site is the official SORA circulating figure, not on-chain total issuance.\n\nRate limits are per IP and per route (10–300 requests per minute); a 429 means wait for the next minute. Agents should prefer the MCP endpoint at `/mcp`, which wraps these routes as typed tools.",
  "license": {
   "name": "MIT",
   "identifier": "MIT"
  }
 },
 "servers": [
  {
   "url": "/",
   "description": "This host"
  }
 ],
...
 "paths": {
...
  "/pools": {
   "get": {
    "operationId": "listPools",
    "summary": "Liquidity pools",
    "description": "XYK pools ordered by value locked; reserves are raw integers (18 decimals) with thousands separators. Illiquid assets are priced 0.",
    "tags": [
     "markets"
    ],
    "responses": {
     "200": {
      "description": "OK",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "description": "Paginated envelope.",
         "properties": {
          "data": {
           "type": "array",
           "items": {
            "type": "object",
            "additionalProperties": true
           }
          },
          "total": {
           "type": "integer"
          },
          "page": {
           "type": "integer"
          },
          "totalPages": {
           "type": "integer"
          }
         },
         "required": [
          "data",
          "total",
          "page",
          "totalPages"
         ]
        }
       }
      }
     },
     "400": {
      "description": "Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Error"
        }
       }
      }
     },
     "429": {
      "description": "Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Error"
        }
       }
      }
     }
    },
    "parameters": [
     {
      "name": "page",
      "in": "query",
      "description": "1-based page. Out-of-range pages are clamped to the last one.",
      "schema": {
       "type": "integer",
       "minimum": 1,
       "default": 1
      }
     },
     {
      "name": "limit",
      "in": "query",
      "description": "Rows per page.",
      "schema": {
       "type": "integer",
       "minimum": 1,
       "maximum": 100,
       "default": 25
      }
     },
     {
      "name": "base",
      "in": "query",
      "description": "Base asset symbol (XOR, XSTUSD, KUSD…).",
      "schema": {
       "type": "string"
      }
     }
    ]
   }
  },
...
  "/search": {
   "get": {
    "operationId": "search",
    "summary": "Search",
    "description": "Resolves an address, block number, extrinsic id (`block-index`), transaction hash or token symbol.",
    "tags": [
     "chain"
    ],
    "responses": {
     "200": {
      "description": "OK",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true
        }
       }
      }
     },
     "400": {
      "description": "Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Error"
        }
       }
      }
     },
     "429": {
      "description": "Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Error"
        }
       }
      }
     }
    },
    "parameters": [
     {
      "name": "q",
      "in": "query",
      "required": true,
      "schema": {
       "type": "string",
       "minLength": 3,
       "maxLength": 128
      }
     }
    ]
   }
  },
  "/holders/{assetId}": {
   "get": {
    "operationId": "listHolders",
    "summary": "Top holders of an asset",
    "description": "25 per page, ranked by free balance read from the chain. `totalHolders` counts accounts above a threshold (>1 XOR, >0.1 other assets), not every account.",
    "tags": [
     "chain"
    ],
    "responses": {
     "200": {
      "description": "OK",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true
        }
       }
      }
     },
     "400": {
      "description": "Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Error"
        }
       }
      }
     },
     "429": {
      "description": "Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Error"
        }
       }
      }
     },
     "503": {
      "description": "A full chain scan is still running; retry after the `Retry-After` header (seconds).",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Error"
        }
       }
      }
     }
    },
    "parameters": [
     {
      "name": "assetId",
      "in": "path",
      "required": true,
      "description": "0x-prefixed 32-byte asset id (from /tokens).",
      "schema": {
       "type": "string",
       "pattern": "^0x[0-9a-fA-F]{64}$"
      }
     },
     {
      "name": "page",
      "in": "query",
      "description": "1-based page. Out-of-range pages are clamped to the last one.",
      "schema": {
       "type": "integer",
       "minimum": 1,
       "default": 1
      }
     }
    ]
   }
  },
  "/burns/stats/{symbol}": {
   "get": {
    "operationId": "getBurnStats",
    "summary": "Burn statistics",
    "description": "Supply and burn figures per window (24h, 7d, 30d, all). For XOR the windows carry the measured supply drop, the fee-mechanism burn and explicit burns. The `all` window spans the 2026 redenomination and is indicative only.",
    "tags": [
     "markets"
    ],
    "responses": {
     "200": {
      "description": "OK",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true
        }
       }
      }
     },
     "400": {
      "description": "Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Error"
        }
       }
      }
     },
     "429": {
      "description": "Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Error"
        }
       }
      }
     }
    },
    "parameters": [
     {
      "name": "symbol",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "enum": [
        "XOR",
        "VAL",
        "PSWAP",
        "TBCD",
        "KUSD"
       ]
      }
     }
    ]
   }
  },
...
 },
 "components": {
  "schemas": {
   "Error": {
    "type": "object",
    "properties": {
     "error": {
      "type": "string"
     },
     "code": {
      "type": "string"
     },
     "message": {
      "type": "string"
     }
    },
    "required": [
     "error"
    ]
   }
  }
 }
}