# crates/api/src/state.rs
//! Shared application state passed to every handler via axum's
//! [`State`](axum::extract::State) extractor.

use crate::chain::ChainClient;
use chrono_tz::Tz;
use sorametrics_db::sm::{load_asset_registry, RegistryAsset};
use sorametrics_db::DbError;
use sorametrics_iroha::ToriiClient;
use sqlx::PgPool;
use std::collections::HashMap;
use std::sync::Arc;
use std::time::Duration;
use tokio::sync::{Mutex, RwLock};
use tracing::{info, warn};

/// Time zone the legacy Node rendered `time` strings in (its process
/// ran with the VPS clock, Europe/Madrid). The frontend parses that
/// string as browser-local time, so the API must keep producing the
/// same wall-clock text. Override with `API_TIME_ZONE`.
pub const DEFAULT_TIME_ZONE: Tz = chrono_tz::Europe::Madrid;

/// How often the in-memory asset registry is reloaded from the DB.
pub const REGISTRY_REFRESH: Duration = Duration::from_secs(300);

/// In-memory snapshot of `sm.asset_registry`, indexed both ways.
#[derive(Default)]
pub struct Registry {
    by_id: HashMap<String, RegistryAsset>,
    /// Canonical asset per symbol: a whitelisted one when it exists,
    /// else the lowest asset id. The legacy registry carries duplicate
    /// symbols (XOR ×8, ETH ×4); the Node's map was "last row wins",
    /// which is not deterministic — this is.
    by_symbol: HashMap<String, String>,
}

impl Registry {
    /// Build from rows.
    pub fn from_rows(rows: Vec<RegistryAsset>) -> Self {
        let mut by_symbol: HashMap<String, String> = HashMap::new();
        let mut sorted: Vec<&RegistryAsset> = rows.iter().collect();
        sorted.sort_by(|a, b| {
            b.whitelisted
                .cmp(&a.whitelisted)
                .then_with(|| a.asset_id.cmp(&b.asset_id))
        });
        for a in sorted {
            by_symbol
                .entry(a.symbol.clone())
                .or_insert_with(|| a.asset_id.clone());
        }
        Self {
            by_id: rows.into_iter().map(|a| (a.asset_id.clone(), a)).collect(),
            by_symbol,
        }
    }

    /// Canonical asset id for a symbol (exact, case-sensitive like the
    /// Node's `_symbolToAssetId`).
    pub fn asset_id_for_symbol(&self, symbol: &str) -> Option<&str> {
        self.by_symbol.get(symbol).map(String::as_str)
    }

    /// `(symbol, canonical asset id)` for every symbol containing
    /// `needle` (case-insensitive), sorted by symbol (Node `/search`).
    pub fn symbols_matching(&self, needle: &str) -> Vec<(String, String)> {
        let up = needle.to_uppercase();
        let mut out: Vec<(String, String)> = self
            .by_symbol
            .iter()
            .filter(|(sym, _)| sym.to_uppercase().contains(&up))
            .map(|(s, id)| (s.clone(), id.clone()))
            .collect();
        // Exact symbol first, then alphabetical.
        out.sort_by(|a, b| {
            (a.0.to_uppercase() != up)
                .cmp(&(b.0.to_uppercase() != up))
                .then_with(|| a.0.cmp(&b.0))
        });
        out
    }

    /// The whitelisted assets (the Node's `ASSETS`), unordered.
    pub fn whitelisted(&self) -> Vec<&RegistryAsset> {
        self.by_id.values().filter(|a| a.whitelisted).collect()
    }

    /// Registry entry for an asset id, if listed.
    pub fn get(&self, asset_id: &str) -> Option<&RegistryAsset> {
        self.by_id.get(asset_id)
    }

    /// Asset ids whose symbol contains `needle` (case-insensitive) — the
    /// legacy `UPPER(symbol) LIKE '%F%'` filter, resolved to ids.
    pub fn asset_ids_matching(&self, needle: &str) -> Vec<String> {
        let up = needle.to_uppercase();
        let mut ids: Vec<String> = self
            .by_id
            .values()
            .filter(|a| a.symbol.to_uppercase().contains(&up))
            .map(|a| a.asset_id.clone())
            .collect();
        ids.sort();
        ids
    }

    /// Number of assets loaded.
    pub fn len(&self) -> usize {
        self.by_id.len()
    }

    /// `true` when nothing is loaded.
    pub fn is_empty(&self) -> bool {
        self.by_id.is_empty()
    }
}

/// Cheap-to-clone state container (pool + `Arc`s).
#[derive(Clone)]
pub struct AppState {
    /// PostgreSQL connection pool.
    pub db: PgPool,
    /// Pool for listings with optional filters (`$n IS NULL OR col = $n`):
    /// its sessions plan every execution with the actual parameters, which
    /// a cached generic plan cannot do. Same database as `db`.
    pub listing_db: PgPool,
    /// Asset registry snapshot, refreshed by [`AppState::spawn_registry_refresh`].
    pub registry: Arc<RwLock<Registry>>,
    /// Zone for legacy `time` strings.
    pub time_zone: Tz,
    /// Read-only chain access for group D routes; `None` = not configured.
    pub chain: Option<ChainClient>,
    /// Archive node for reads at historical blocks (the Node's
    /// `getArchiveApi()`); `None` = not configured.
    pub archive: Option<ChainClient>,
    /// Minamoto Torii client for `/api/minamoto/*`; `None` = not configured.
    pub torii: Option<ToriiClient>,
    /// Process start, for `/health.uptime`.
    pub started_at: std::time::Instant,
    /// In-process TTL caches for the heavy chain scans (`/pools`,
    /// `/holders`), keyed by route-specific strings.
    pub scan_cache: Arc<Mutex<HashMap<String, CachedScan>>>,
    /// Scan keys currently being refreshed in the background.
    pub scans_in_flight: Arc<Mutex<std::collections::HashSet<String>>>,
    /// Site analytics queue, presence and caches (`/analytics/*`).
    pub analytics: Arc<crate::routes::analytics::Analytics>,
    /// Assets whose holders were requested, by last request time: the
    /// pre-warm loop keeps their scans fresh.
    pub hot_holders: Arc<Mutex<HashMap<String, std::time::Instant>>>,
}

/// One cached scan result with its expiry.
#[derive(Clone)]
pub struct CachedScan {
    /// When the entry was produced.
    pub at: std::time::Instant,
    /// JSON-serialisable payload (already shaped for the route).
    pub value: serde_json::Value,
}

impl AppState {
    /// Cached scan value if younger than `ttl`.
    pub async fn cached_scan(&self, key: &str, ttl: Duration) -> Option<serde_json::Value> {
        let cache = self.scan_cache.lock().await;
        cache
            .get(key)
            .filter(|c| c.at.elapsed() < ttl)
            .map(|c| c.value.clone())
    }

    /// Store a scan result.
    pub async fn store_scan(&self, key: &str, value: serde_json::Value) {
        self.scan_cache.lock().await.insert(
            key.to_string(),
            CachedScan {
                at: std::time::Instant::now(),
                value,
            },
        );
    }

    /// Mark a scan as running; `false` if it already was.
    pub async fn begin_scan(&self, key: &str) -> bool {
        self.scans_in_flight.lock().await.insert(key.to_string())
    }

    /// Clear the running mark.
    pub async fn end_scan(&self, key: &str) {
        self.scans_in_flight.lock().await.remove(key);
    }
}

impl AppState {
    /// Constructs from a live pool with an EMPTY registry (tests, or
    /// callers that will load it themselves).
    pub fn new(db: PgPool) -> Self {
        Self {
            listing_db: db.clone(),
            db,
            registry: Arc::new(RwLock::new(Registry::default())),
            time_zone: DEFAULT_TIME_ZONE,
            chain: None,
            archive: None,
            torii: None,
            started_at: std::time::Instant::now(),
            scan_cache: Arc::new(Mutex::new(HashMap::new())),
            scans_in_flight: Arc::new(Mutex::new(std::collections::HashSet::new())),
            analytics: Arc::new(crate::routes::analytics::Analytics::from_env()),
            hot_holders: Arc::new(Mutex::new(HashMap::new())),
        }
    }

    /// Use a separate pool for filtered listings (see `listing_db`).
    pub fn with_listing_db(mut self, pool: PgPool) -> Self {
        self.listing_db = pool;
        self
    }

    /// Attach a chain client.
    pub fn with_chain(mut self, chain: Option<ChainClient>) -> Self {
        self.chain = chain;
        self
    }

    /// Attach the archive chain client.
    pub fn with_archive(mut self, archive: Option<ChainClient>) -> Self {
        self.archive = archive;
        self
    }

    /// Attach a Torii client.
    pub fn with_torii(mut self, torii: Option<ToriiClient>) -> Self {
        self.torii = torii;
        self
    }

    /// Loads the registry once from the DB. Fails loudly: an API with
    /// no symbols would render every row as `0xXXXX`.
    pub async fn with_registry(db: PgPool, time_zone: Tz) -> Result<Self, DbError> {
        let rows = load_asset_registry(&db).await?;
        info!(assets = rows.len(), zone = %time_zone, "asset registry loaded");
        Ok(Self {
            listing_db: db.clone(),
            db,
            registry: Arc::new(RwLock::new(Registry::from_rows(rows))),
            time_zone,
            chain: None,
            archive: None,
            torii: None,
            started_at: std::time::Instant::now(),
            scan_cache: Arc::new(Mutex::new(HashMap::new())),
            scans_in_flight: Arc::new(Mutex::new(std::collections::HashSet::new())),
            analytics: Arc::new(crate::routes::analytics::Analytics::from_env()),
            hot_holders: Arc::new(Mutex::new(HashMap::new())),
        })
    }

    /// Record a `/holders` request for `asset_id`.
    pub async fn note_hot_holder(&self, asset_id: &str) {
        self.hot_holders
            .lock()
            .await
            .insert(asset_id.to_string(), std::time::Instant::now());
    }

    /// Assets requested within `window`, most recent first, at most `cap`.
    pub async fn hot_holders(&self, window: Duration, cap: usize) -> Vec<String> {
        let mut hot = self.hot_holders.lock().await;
        hot.retain(|_, at| at.elapsed() <= window);
        let mut v: Vec<(&String, &std::time::Instant)> = hot.iter().collect();
        v.sort_by(|a, b| b.1.cmp(a.1));
        v.into_iter().take(cap).map(|(a, _)| a.clone()).collect()
    }

    /// Reloads the registry from the DB every [`REGISTRY_REFRESH`].
    /// A failed reload keeps the previous snapshot and logs.
    pub fn spawn_registry_refresh(&self) {
        let db = self.db.clone();
        let registry = self.registry.clone();
        tokio::spawn(async move {
            let mut ticker = tokio::time::interval(REGISTRY_REFRESH);
            ticker.tick().await;
            loop {
                ticker.tick().await;
                match load_asset_registry(&db).await {
                    Ok(rows) => {
                        *registry.write().await = Registry::from_rows(rows);
                    }
                    Err(e) => warn!(error = %e, "asset registry refresh failed; keeping previous"),
                }
            }
        });
    }
}

/// Parses `API_TIME_ZONE` (IANA name) or falls back to [`DEFAULT_TIME_ZONE`].
pub fn time_zone_from_env() -> Result<Tz, String> {
    match std::env::var("API_TIME_ZONE") {
        Ok(v) => v
            .parse::<Tz>()
            .map_err(|e| format!("API_TIME_ZONE '{v}' is not an IANA zone: {e}")),
        Err(_) => Ok(DEFAULT_TIME_ZONE),
    }
}

#[cfg(test)]
mod tests {
    use super::*;

    fn asset(id: &str, sym: &str) -> RegistryAsset {
        RegistryAsset {
            asset_id: id.to_string(),
            symbol: sym.to_string(),
            name: None,
            decimals: 18,
            logo: None,
            whitelisted: true,
        }
    }

    #[test]
    fn symbol_resolves_to_whitelisted_then_lowest_id() {
        let mut dup = asset("0x09", "XOR");
        dup.whitelisted = false;
        let r = Registry::from_rows(vec![dup, asset("0x02", "XOR"), asset("0x01", "XOR")]);
        assert_eq!(r.asset_id_for_symbol("XOR"), Some("0x01"));
        let mut a = asset("0x05", "ETH");
        a.whitelisted = false;
        let mut b = asset("0x03", "ETH");
        b.whitelisted = false;
        let r = Registry::from_rows(vec![a, b]);
        assert_eq!(r.asset_id_for_symbol("ETH"), Some("0x03"));
        assert_eq!(r.asset_id_for_symbol("eth"), None);
    }

    #[test]
    fn symbol_filter_is_case_insensitive_substring() {
        let r = Registry::from_rows(vec![
            asset("0x01", "XOR"),
            asset("0x02", "KXOR"),
            asset("0x03", "DAI"),
        ]);
        assert_eq!(r.asset_ids_matching("xor"), vec!["0x01", "0x02"]);
        assert_eq!(r.asset_ids_matching("DAI"), vec!["0x03"]);
        assert!(r.asset_ids_matching("zzz").is_empty());
    }
}